fix(editor): removed fullscreen functionality
The full screen functionality in the editor didn't work (produced weird result). There currently seems to be no way to make fullscreen work in an easy way, as the H5P editor client hides all other elements except the editor iframe through JavaScript. This means all the Chrome provided by React disappears and we'd have to change a lot of styling for this to work. I've simply opted for removing the button and the disabling the fullscreen functionality.