fix(user interface): improvements in open file list
The file list now displays filenames of newly saved files. It now also only displays the filename and not the full path on Windows. This was caused by Windows path names with Windows path separaors (\) being used in the Electron window, which expects POSIX separators (/).

I've also added an ellipsis and a tooltip to the tab items.