refactor(error-handler): pipe error-responses through error-middleware
This change makes the server conform with the express-style to handle errrors: https://expressjs.com/en/guide/error-handling.html
It adds a central errorHandler middleware at the end of the middleware-chain, which is called when a handler calls the next(error) function with an argument. (Usually an error)
This allows us to remove quite a lot sentry imports and it provides a central place to send back an error.