refactor: client doesn't use paths when opening or saving
The old way of saving and loading h5ps was insecure, as it
would have been possible for a malicious H5P library to send client
requests that do something undesired on the user's disk. Now there are
only file handles that reference files the user has picked in the file
picker before.